+++ app/builders/ndr_ui/bootstrap/readonly.rb
@@ -10,7 +10,7 @@ module NdrUi
       def self.included(base)
         # These have different signatures, or aren't affected by `readonly`:
         not_affected = [:label, :fields_for]
-        needs_custom = [:radio_button, :file_field, :hidden_field] +
+        needs_custom = [:radio_button, :file_field, :hidden_field, :textarea] +
                        base.field_helpers_from_form_options_helper
 
         (base.field_helpers - needs_custom - not_affected).each do |selector|
@@ -19,6 +19,16 @@ module NdrUi
               return super unless readonly?
               readonly_value = options.symbolize_keys.fetch(:readonly_value, object.send(method))
               @template.content_tag(:p, readonly_value, class: 'form-control-static')
+            end
+          END
+        end
+
+        [:text_area].each do |selector|
+          class_eval <<-END, __FILE__, __LINE__ + 1
+            def #{selector}(method, options = {}, *rest)
+              return super unless readonly?
+              readonly_value = options.symbolize_keys.fetch(:readonly_value, object.send(method))
+              @template.simple_format(readonly_value, class: 'form-control-static')
             end
           END
         end